Fix tensor products with trivial factors
Fixes a bug where tensor products with Z and I factors would fail
because the simulator would attempy to apply  their matrices, which act
only on a subset of the target, to the entire target. This is done by
overhauling the way trivial operations are treated.

As well, the name "constituent" is changed to the more proper "factor."