feature: Calculate arbitrary observables (#75)
Allows arbitrary observables to be attached to a circuit, allowing, for example, the simultaneous calculation of all terms of a Hamiltonian.

This change only applies for `shots=0`, as result types are only defined for `shots=0` in this simulator. The `shots>0` case is handled entirely by the SDK, using samples generated by the simulator.